+++ src/main/java/blue/steel/backend/story/campaign/adapter/CampaignQueryController.java
@@ -1,7 +1,7 @@
 package blue.steel.backend.story.campaign.adapter;
 
 import blue.steel.backend.story.campaign.entity.Campaign;
-import blue.steel.backend.story.campaign.usecase.CampaignQueryService;
+import blue.steel.backend.story.campaign.usecase.CampaignQuery;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
 
@@ -9,14 +9,14 @@ import org.springframework.stereotype.Controller;
 @Controller
 public class CampaignQueryController {
 
-  private final CampaignQueryService campaignQueryService;
+  private final CampaignQuery campaignQuery;
 
-  public CampaignQueryController(CampaignQueryService campaignService) {
-    this.campaignQueryService = campaignService;
+  public CampaignQueryController(CampaignQuery campaignQuery) {
+    this.campaignQuery = campaignQuery;
   }
 
   @QueryMapping
   public Campaign actualCampaign() {
-    return campaignQueryService.findActualCampaign();
+    return campaignQuery.findActualCampaign();
   }
 }
