+++ src/test/java/blue/steel/backend/story/campaign/integration/CreateCampaignUseCaseTest.java
@@ -3,11 +3,7 @@ package blue.steel.backend.story.campaign.integration;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import blue.steel.backend.story.campaign.adapter.dto.CreateCampaignInput;
-import blue.steel.backend.story.campaign.adapter.dto.UpdateCampaignInput;
 import blue.steel.backend.story.campaign.entity.Campaign;
-import blue.steel.backend.story.campaign.entity.CampaignRepository;
-import blue.steel.backend.story.campaign.entity.CampaignRepositoryTest;
-import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,13 +11,12 @@ import org.springframework.boot.test.autoconfigure.graphql.tester.AutoConfigureW
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.graphql.test.tester.WebGraphQlTester;
 
-/** Campaign mutation controller tests. */
+/** Create campaign use case tests. */
 @SpringBootTest
 @AutoConfigureWebGraphQlTester
-public class CampaignMutationControllerTest {
+public class CreateCampaignUseCaseTest {
 
   @Autowired private WebGraphQlTester graphQlTester;
-  @Autowired private CampaignRepository campaignRepository;
 
   @Test
   @DisplayName("Creating a valid campaign should return a not null campaign")
@@ -40,29 +35,5 @@ public class CampaignMutationControllerTest {
 
         // Then response should contain a campaign
         .satisfies(campaign -> assertThat(campaign).isNotNull());
-  }
-
-  @Test
-  @DisplayName("Updating a valid campaign should return a not null campaign")
-  void updateValidCampaign() {
-    // Given an existing campaign
-    Campaign campaign = CampaignRepositoryTest.createCampaign();
-    campaign = campaignRepository.save(campaign);
-
-    // And a valid update campaign input
-    UUID id = campaign.getId();
-    UpdateCampaignInput createCampaignInput =
-        new UpdateCampaignInput(id, "new name", "new description", "new imageUrl");
-
-    // When updating a campaign
-    graphQlTester
-        .queryName("story/campaign/queries/updateCampaign")
-        .variable("input", createCampaignInput)
-        .execute()
-        .path("updateCampaign.campaign")
-        .entity(Campaign.class)
-
-        // Then response should contain a campaign
-        .satisfies(updatedCampaign -> assertThat(updatedCampaign).isNotNull());
   }
 }
